node_package/requirements.txt
# DecentraStore Node Dependencies
python-socketio[client]>=5.0.0
websocket-client>=1.0.0
